refactor: move deliverCargo() method to separate module
